backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.55.3
httpx>=0.28.1
typing-extensions>=4.14.0